LayoutTests/fast/js/dfg-store-unexpected-value-into-argument-and-osr-exit-expected.txt

Tests what happens when you store an unexpected value into an argument (where the original argument value was the expected one) and then OSR exit.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 1
PASS foo(3, {f:f, g:g}, p) is 85
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS foo(3, {f:f, g:g}, p) is 0.5
PASS foo(3, {f:f, g:g}, p) is 85.5
PASS successfullyParsed is true

TEST COMPLETE
